fix webpack config for css
