+++ package.json
@@ -61,6 +61,7 @@
     "typescript": "*",
     "webpack": "^5.61.0",
     "webpack-cli": "^5.0.1",
-    "webpack-merge": "^5.0.0"
+    "webpack-merge": "^5.0.0",
+    "webpack-remove-empty-scripts": "^1.0.4"
   }
 }
