+++ src/static/manifest.json
@@ -53,6 +53,36 @@
         "*://schoology.com/*",
         "*://lti-submission-google.app.schoology.com/*"
       ],
+      "css": [
+        "styles/all.css",
+        "styles/modern/all.css",
+        "styles/lib/contextmenu.css",
+        "styles/lib/izitoast.css",
+        "styles/lib/jquery-ui.css"
+      ],
+      "run_at": "document_start"
+    },
+    {
+      "matches": [
+        "https://lms.lausd.net/*",
+        "https://*.schoology.com/*"
+      ],
+      "exclude_matches": [
+        "https://*.schoology.com/login*",
+        "https://*.schoology.com/register*",
+        "*://asset-cdn.schoology.com/*",
+        "*://developer.schoology.com/*",
+        "*://support.schoology.com/*",
+        "*://info.schoology.com/*",
+        "*://files-cdn.schoology.com/*",
+        "*://status.schoology.com/*",
+        "*://ui.schoology.com/*",
+        "*://www.schoology.com/*",
+        "*://api.schoology.com/*",
+        "*://developers.schoology.com/*",
+        "*://schoology.com/*",
+        "*://lti-submission-google.app.schoology.com/*"
+      ],
       "js": [
         "content.js"
       ],
