+++ webpack.config.js
@@ -1,15 +1,31 @@
 const { merge } = require("webpack-merge");
 const path = require("path");
 const CopyPlugin = require("copy-webpack-plugin");
+const MiniCssExtractPlugin = require("mini-css-extract-plugin");
+const RemoveEmptyScriptsPlugin = require("webpack-remove-empty-scripts");
 const srcDir = path.join(__dirname, ".", "src");
 
 const commonConfig = {
     entry: {
+        // Typescript Entry Points
         popup: path.join(srcDir, "scripts/popup.ts"),
         options: path.join(srcDir, "scripts/options.ts"),
         background: path.join(srcDir, "scripts/background.ts"),
         content: path.join(srcDir, "scripts/content.ts"),
         offscreen: path.join(srcDir, "scripts/offscreen.ts"),
+
+        // CSS Entry Points
+        "styles/all": path.join(srcDir, "styles/all.scss"),
+        "styles/modern/all": path.join(srcDir, "styles/modern/all.scss"),
+        "styles/lib/jquery-ui": path.join(srcDir, "../node_modules/jquery-ui/themes/base/all.css"),
+        "styles/lib/contextmenu": path.join(
+            srcDir,
+            "../node_modules/jquery-contextmenu/dist/jquery.contextMenu.min.css"
+        ),
+        "styles/lib/izitoast": path.join(
+            srcDir,
+            "../node_modules/iziToast/dist/css/iziToast.min.css"
+        ),
     },
     output: {
         filename: "[name].js",
@@ -25,8 +41,9 @@ const commonConfig = {
             {
                 test: /\.(scss|css)$/,
                 use: [
+                    MiniCssExtractPlugin.loader,
                     // Creates `style` nodes from JS strings
-                    "style-loader",
+                    // "style-loader",
                     // Translates CSS into CommonJS (for postcss-loader because it takes css as input)
                     {
                         loader: "css-loader",
@@ -57,6 +74,10 @@ const commonConfig = {
                 { from: path.resolve("src/html"), to: path.resolve("dist") },
             ],
             options: {},
+        }),
+        new RemoveEmptyScriptsPlugin(),
+        new MiniCssExtractPlugin({
+            filename: "[name].css",
         }),
     ],
 };
